Updated to help with redundancy
I remembered that you can't download the default images (the cat and dog ones) so I will work on adding the option to access them online later
Made the pizza-paper.sh check to see that you have $/Documents/pizzapapers.txt file, and if you don't it will add it  (including the dir if you don't already)
Added debugging stuff to path-adder.sh so that I don't have to copy a new version of my development file as pizza-share.sh to test it

+++ path-adder.sh
@@ -50,7 +50,11 @@ ResetSettings (){
 options=$(getopt -o ar,add,remove --long "add,remove,retard,copy" -- "$@")
 =======
   if ! test -f /usr/local/bin/pizzapaper; then		#If pizzapaper is NOT found in the path file, it will add it
-    cp ./pizza-paper.sh /usr/local/bin/pizzapaper
+    if test -f */pizzapaper; then
+      cp ./pizza-paper.sh /usr/local/bin/pizzapaper
+    else
+      cp ./pizzapaper_testing /usr/local/bin/pizzapaper
+    fi
   else
     echo "pizzapaper is already in the path directory: /usr/local/bin/"
   fi
